requirements.txt
streamlit
ultralytics
torch
torchvision
Pillow
numpy